int for byte

+++ type-variables/src/type_variables.java
@@ -4,7 +4,7 @@ public class type_variables {
         long big_number = 1032302L;
         short small_number = 1; 
         double money = 2.453;
-        int age = 10;
+        byte age = 10;
         float pi  = 3.14F;
 
         System.out.println();
